requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
dnspython==2.7.0
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
fsspec==2025.2.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
numpy==2.2.3
packaging==24.2
portalocker==2.10.1
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pyobjectid==0.1.3
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.13.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
tavily-python==0.5.1
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
